coord,dataflow: move all common files to one place
